Implement MindboxHuawei.handleRemoteMessage
