+++ mindbox-hms/build.gradle
@@ -38,4 +38,5 @@ dependencies {
     implementation 'com.huawei.hms:push:6.1.0.300'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.2'
     implementation 'com.huawei.hms:ads-identifier:3.4.39.302'
+    implementation 'androidx.annotation:annotation:1.3.0'
 }
