+++ mindbox-hms/src/main/java/cloud/mindbox/mindbox_hms/HuaweiRemoteMessageTransformer.kt
@@ -1,0 +1,10 @@
+package cloud.mindbox.mindbox_hms
+
+import com.huawei.hms.push.RemoteMessage as HuaweiRemoteMessage
+import cloud.mindbox.mobile_sdk_core.pushes.RemoteMessage
+
+internal object HuaweiRemoteMessageTransformer {
+
+    fun transform(remoteMessage: HuaweiRemoteMessage?): RemoteMessage? = TODO("Not implemented")
+
+}
